compilation ok. not perfect, prompt missing part1

+++ include/unistd.h
@@ -32,5 +32,5 @@ int pipe(int pipefd[2]);
 
 //functions add
 
-int write(int fd, const void *buf, size_t count);
+ssize_t write(int fd, const void *buf, size_t count);
 #endif
